+++ q2_types/feature_table/__init__.py
@@ -1,0 +1,17 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2016--, QIIME development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+
+import importlib
+
+from ._format import BIOMV1Format, FeatureTableDirectoryFormat
+from ._type import FeatureTable, Frequency, RelativeFrequency, PresenceAbsence
+
+__all__ = ['BIOMV1Format', 'FeatureTableDirectoryFormat', 'FeatureTable',
+           'Frequency', 'RelativeFrequency', 'PresenceAbsence']
+
+importlib.import_module('q2_types.feature_table._transformer')
